% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-varnames.R
\name{abbreviate_v}
\alias{abbreviate_v}
\title{Vectorized Abbreviation}
\usage{
abbreviate_v(names.arg, minlength)
}
\arguments{
\item{names.arg}{a character vector of names to be abbreviated, or an object
to be coerced to a character vector by as.character.}

\item{minlength}{the minimum length of the abbreviations, as a numeric
vector. Must be same length as names.arg argument's vector.}
}
\description{
Makes the \code{abbreviate} function from base R vectorized to
accommodate a vector of differing minLength values. Cannot use
\code{Vectorize} here, as it will lead to dupes. This method
generates the abbreviations with the largest minLengths first and leaves
those intact as it continues onto the terms with the next highest minLength
values. This continues iteratively until the lowest minlength is reached.
}
\examples{
abbreviate_v(c("Fruit Flies", "Tubular Dude"),c(4,6))
}
