% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{add_label}
\alias{add_label}
\title{Add label to a vector}
\usage{
add_label(x, label)
}
\arguments{
\item{.data}{Dataframe with labels to extract}
}
\value{
A vector of variable and labels for a dataframe
}
\description{
Add label to a vector
}
\keyword{internal}
