% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to a Tibble}
\usage{
add_labels(.data, ...)
}
\arguments{
\item{.data}{A tibble or dataframe that are in need of labels}

\item{...}{additional abilities as needed}
}
\value{
A vector of variable and labels for a dataframe
}
\description{
Add labels to a Tibble
}
