% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{check_ascii_lbls}
\alias{check_ascii_lbls}
\title{ASCII Check on Variable Labels}
\usage{
check_ascii_lbls(.data)
}
\value{

}
\description{
Check that only ASCII characters are being used in variable labels
}
