% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{check_ascii_vars}
\alias{check_ascii_vars}
\title{ASCII Check on Variable Names}
\usage{
check_ascii_vars(.data)
}
\value{

}
\description{
Check that only ASCII characters are being used in variable names
}
