% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{check_label_length}
\alias{check_label_length}
\title{Variable Label Check}
\usage{
check_label_length(.data)
}
\arguments{
\item{.data}{Data set with labels to check}
}
\value{

}
\description{
Check for variable labels greater than 40 character length
}
