% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{check_var_case}
\alias{check_var_case}
\title{UPPER CASE check}
\usage{
check_var_case(.data)
}
\arguments{
\item{.data}{Data frame containing data to calculate summary statistics for}
}
\value{
Tibble of Variables flagged with lower case
}
\description{
Check to see if data set has all upper case for variables
}
