% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-varnames.R
\name{chg_letter_case}
\alias{chg_letter_case}
\title{Change letter case - upper, lower, or leave as-is}
\usage{
chg_letter_case(x, letter_case = "asis")
}
\arguments{
\item{x}{a character vector, or an object which can be coerced by
as.character to a character vector. Long vectors are supported.}

\item{letter_case}{character string, either "lower", "upper", or "asis"}
}
\description{
Supply a character string or vector, returns the same vector with changed
case. Notice, this function just uses \code{toupper()} or \code{tolower()} but the
string, \code{x}, may have been altered by the janitor package's use of
\code{\link[snakecase]{to_any_case}} later on, when suggesting a rename
}
\examples{
chg_letter_case(c("hello darkness","My Old FRIEND"), "lower" )
chg_letter_case(c("hello darkness","My Old FRIEND"), "upper" )
chg_letter_case(c("hello darkness","My Old FRIEND"), "asis" )
}
