% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\docType{data}
\name{cli_theme_tests}
\alias{cli_theme_tests}
\title{Theme for cli package messages when running inside tests}
\format{
An object of class \code{list} of length 1.
}
\usage{
cli_theme_tests
}
\description{
It can be defined with starting an \code{cli::start_app(theme = cli_theme_tests)}
}
\keyword{datasets}
