% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{define_dflabel}
\alias{define_dflabel}
\title{Dataset label}
\usage{
define_dflabel(x, y)
}
\arguments{
\item{x}{A data frame of CDISC standard.}

\item{y}{A data frame containing dataset level metadata.}
}
\value{
Data frame with label attributes.
}
\description{
Assigns dataset label from a dataset level metadata to a given data frame.
}
\examples{
adsl <- tibble::tribble(
 ~USUBJID, ~SITEID, ~AGE, ~SEX,
 1001    , 001    , 11  , "M",
 1002    , 002    , 10  , "F",
 1003    , 003    , 14  , "M"
)

dsmeta <- tibble::tribble(
 ~name , ~label,
 "adsl", "Subject-Level Analysis",
 "adae", "Adverse Events Analysis"
)

adsl <- define_dflabel(adsl, dsmeta)
}
\seealso{
\code{\link[=define_varlabel]{define_varlabel()}} and \code{\link[=define_format]{define_format()}}

Other metadata functions: 
\code{\link{define_format}()},
\code{\link{define_varlabel}()}
}
\concept{metadata functions}
