% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{define_format}
\alias{define_format}
\title{SAS format}
\usage{
define_format(x, y)
}
\arguments{
\item{x}{A data frame of CDISC standard.}

\item{y}{A data frame containing variable level metadata.}
}
\value{
Data frame with \code{SASformat} attributes for each variable.
}
\description{
Assigns SAS format from a variable level metadata to a given data frame.
}
\examples{
adsl <- tibble::tribble(
 ~USUBJID, ~BRTHDT,
 1001    , 1,
 1002    , 1,
 1003    , 2      
)

metadata <- tibble::tribble(
 ~dataset, ~variable, ~sas_format,
 "adsl"  , "USUBJID", NA,
 "adsl"  , "BRTHDT" , "date9."
 )

adsl <- define_format(adsl, metadata)
}
\seealso{
\code{\link[=define_varlabel]{define_varlabel()}} and \code{\link[=define_dflabel]{define_dflabel()}}

Other metadata functions: 
\code{\link{define_dflabel}()},
\code{\link{define_varlabel}()}
}
\concept{metadata functions}
