% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{define_varlabel}
\alias{define_varlabel}
\title{Variable label}
\usage{
define_varlabel(x, y)
}
\arguments{
\item{x}{A data frame of CDISC standard.}

\item{y}{A data frame containing variable level metadata.}
}
\value{
Data frame with label attributes for each variable.
}
\description{
Assigns variable label from a variable level metadata to a given data frame.
}
\examples{
adsl <- tibble::tribble(
 ~USUBJID, ~SITEID, ~AGE, ~SEX,
 1001    , 001    , 11  , "M",
 1002    , 002    , 10  , "F",
 1003    , 003    , 14  , "M"
)

varmeta <- tibble::tribble(
 ~dataset, ~variable, ~label,
 "adsl"  , "USUBJID", "Unique Subject Identifier",
 "adsl"  , "SITEID" , "Study Site Identifier",
 "adsl"  , "AGE"    , "Age",
 "adsl"  , "SEX"    , "Sex"
 )

adsl <- define_varlabel(adsl, varmeta)
}
\seealso{
\code{\link[=define_dflabel]{define_dflabel()}} and \code{\link[=define_format]{define_format()}}

Other metadata functions: 
\code{\link{define_dflabel}()},
\code{\link{define_format}()}
}
\concept{metadata functions}
