% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_label.R
\name{df_label_view}
\alias{df_label_view}
\title{View Label on a dataframe}
\usage{
df_label_view(.data)
}
\arguments{
\item{.data}{}
}
\value{
Message with the Label or "No Label on dataframe"
}
\description{
View Label on a dataframe
}
