% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\name{expect_attr_width}
\alias{expect_attr_width}
\title{Custom expect function to test result of xportr_length}
\usage{
expect_attr_width(result, metadata_length)
}
\arguments{
\item{result}{data.frame with \code{width} attribute on its columns.}

\item{metadata_length}{vector of numeric with expected lengths for each
column width.}
}
\value{
The first argument, invisibly.
}
\description{
Custom expect function to test result of xportr_length
}
\keyword{internal}
