% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{extract_label}
\alias{extract_label}
\title{Extract a label from a vector}
\usage{
extract_label(x)
}
\arguments{
\item{x}{A vector with a \"label\" attribute}
}
\value{
The text label associated with the vector or \code{NA}
}
\description{
Extract a label from a vector
}
\keyword{internal}
