% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract all labels from a Dataframe}
\usage{
extract_labels(.data)
}
\arguments{
\item{.data}{Dataframe with labels to extract}
}
\value{
A vector of variable and labels for a dataframe
}
\description{
Extract all labels from a Dataframe
}
\examples{
data(dm)
labs <- extract_labels(dm)
}
