% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{label_log}
\alias{label_log}
\title{Utility for Variable Labels}
\usage{
label_log(miss_vars, verbose)
}
\arguments{
\item{miss_vars}{Missing variables in metadata}

\item{verbose}{Provides additional messaging for user}
}
\value{
Output to Console
}
\description{
Utility for Variable Labels
}
