% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{length_log}
\alias{length_log}
\title{Utility for Lengths}
\usage{
length_log(miss_vars, verbose)
}
\arguments{
\item{miss_vars}{Variables missing from metatdata}

\item{verbose}{Provides additional messaging for user}
}
\value{
Output to Console
}
\description{
Utility for Lengths
}
