% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\name{local_cli_theme}
\alias{local_cli_theme}
\title{Local function to remove extra spaces and format by cli}
\usage{
local_cli_theme(.local_envir = parent.frame())
}
\arguments{
\item{`[environment(1)]`\cr}{Attach exit handlers to this environment. Typically, this should
be either the current environment or a parent frame
(accessed through parent.frame()).}
}
\description{
Groups together multiple calls instead of being spread out in code
}
\keyword{internal}
