% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{max_length_msg}
\alias{max_length_msg}
\title{Utility for data Lengths}
\usage{
max_length_msg(max_length, verbose)
}
\arguments{
\item{max_length}{Dataframe with data and metadata length}

\item{verbose}{Provides additional messaging for user}
}
\value{
Output to Console
}
\description{
Utility for data Lengths
}
