% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\name{multiple_vars_in_spec_helper}
\alias{multiple_vars_in_spec_helper}
\title{Test if multiple vars in spec will result in warning message}
\usage{
multiple_vars_in_spec_helper(fun)
}
\description{
Test if multiple vars in spec will result in warning message
}
\keyword{internal}
