% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{type_log}
\alias{type_log}
\title{Utility for Types}
\usage{
type_log(meta_ordered, type_mismatch_ind, verbose)
}
\arguments{
\item{meta_ordered}{fill in later}

\item{type_mismatch_ind}{fill in later}

\item{verbose}{Provides additional messaging for user}
}
\value{
Output to Console
}
\description{
Utility for Types
}
