% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{var_names_log}
\alias{var_names_log}
\title{Title}
\usage{
var_names_log(tidy_names_df, verbose)
}
\arguments{
\item{tidy_names_df}{fill in later}

\item{verbose}{fill in later}
}
\value{
stuff
}
\description{
Title
}
