% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{var_ord_msg}
\alias{var_ord_msg}
\title{Title}
\usage{
var_ord_msg(moved_vars, verbose)
}
\arguments{
\item{moved_vars}{fill in later}

\item{verbose}{fill in later}
}
\value{
stuff
}
\description{
Title
}
