% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_spec}
\alias{var_spec}
\title{Example Dataset Variable Specification}
\format{
\subsection{\code{var_spec}}{

A data frame with 216 rows and 19 columns:
\describe{
\item{Order}{Order of variable}
\item{Dataset}{Dataset}
\item{Variable}{Variable}
\item{Label}{Variable Label}
\item{Data Type}{Data Type}
\item{Length}{Variable Length}
\item{Significant Digits}{Significant Digits}
\item{Format}{Variable Format}
\item{Mandatory}{Mandatory Variable Flag}
\item{Assigned Value}{Variable Assigned Value}
\item{Codelist}{Variable Codelist}
\item{Common}{Common Variable Flag}
\item{Origin}{Variable Origin}
\item{Pages}{Pages}
\item{Method}{Variable Method}
\item{Predecessor}{Variable Predecessor}
\item{Role}{Variable Role}
\item{Comment}{Comment}
\item{Developer Notes}{Developer Notes}
}
}
}
\usage{
data("var_spec")
}
\description{
Example Dataset Variable Specification
}
\keyword{datasets}
