% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xportr.R
\name{variable_max_length}
\alias{variable_max_length}
\title{Calculate the maximum length of variables}
\usage{
variable_max_length(.df)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}
}
\value{
Returns a dataframe with variables and their maximum length
}
\description{
Function to calculate the maximum length of variables in a given dataframe
}
