% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{xportr_logger}
\alias{xportr_logger}
\title{Title}
\usage{
xportr_logger(message, type = "none", ...)
}
\arguments{
\item{message}{fill in later}

\item{type}{fill in later}

\item{...}{fill in later}
}
\value{
stuff
}
\description{
Title
}
