% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xportr-options.R
\name{xportrOptions}
\alias{xportrOptions}
\title{Get or set Xportr options}
\usage{
xportrOptions(...)
}
\arguments{
\item{...}{Options to set, with the form \code{name = value}.}
}
\description{
There are two mechanisms for working with options for xportr. One is the
\code{\link[=options]{options()}} function, which is part of base R, and the other is the
\code{xportrOptions()} function, which is in the xportr package. The reason for
these two mechanisms is has to do with legacy code and scoping.

The \code{\link[=options]{options()}} function sets options globally, for the duration of the R
process. The \code{\link[=getOption]{getOption()}} function retrieves the value of an option. All
xportr related options of this type are prefixed with \code{"xportr."}.
}
\section{Options with \code{options()}}{


\describe{
\item{xportr.df_domain_name (defaults to \code{"dataset"})}{Description about this option ...}
\item{xportr.df_label (defaults to \code{"label"})}{Description about this option ...}
\item{xportr.domain_name (defaults to \code{"dataset"})}{Description about this option ...}
\item{xportr.variable_name (defaults to \code{"variable"})}{Description about this option ...}
\item{xportr.type_name (defaults to \code{"type"})}{Description about this option ...}
\item{xportr.label (defaults to \code{"label"})}{Description about this option ...}
\item{xportr.length (defaults to \code{"length"})}{Description about this option ...}
\item{xportr.format_name (defaults to \code{"format"})}{Description about this option ...}
\item{xportr.format_verbose (defaults to \code{"none"})}{Description about this option ...}
\item{xportr.label_verbose (defaults to \code{"none"})}{Description about this option ...}
\item{xportr.length_verbose (defaults to \code{"none"})}{Description about this option ...}
\item{xportr.type_verbose (defaults to \code{"label"})}{Description about this option ...}
\item{xportr.character_types (defaults to \code{c("character", "char", "text", "date", "posixct", "posixt", "datetime", "time", "partialdate", "partialtime", "partialdatetime", "incompletedatetime", "durationdatetime", "intervaldatetime")})}{Description about this option ...} # nolint
\item{xportr.numeric_types (defaults to \code{c("integer", "numeric", "num", "float")})}{Description about this option ...} # nolint
\item{xportr.order_name (defaults to \code{"order"})}{Description about this option ...}
}

\section{Options with \code{xportrOptions()}}{


There are a number of global options that affect xportr's behavior. These
can be set globally with \code{options()} or  with \code{xportrOptions()}.
The \code{xportrOptions()} function also returns the current options when nothing is passed to it.
}

\examples{
xportrOptions(xportr.df_label = "data_label", xportr.label = "custom_label")
xportrOptions("xportr.df_label")
xportrOptions(c("xportr.label", "xportr.df_label"))
xportrOptions()
}
