% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{xpt_check_ascii_lbls}
\alias{xpt_check_ascii_lbls}
\title{ASCII Check on Variable Labels}
\usage{
xpt_check_ascii_lbls(.data)
}
\arguments{
\item{.data}{Data frame}
}
\value{

}
\description{
Check that only ASCII characters are being used in variable labels
}
