% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{xpt_check_ascii_vars}
\alias{xpt_check_ascii_vars}
\title{ASCII Check on Variable Names}
\usage{
xpt_check_ascii_vars(.data)
}
\arguments{
\item{.data}{Data frame}
}
\description{
Check that only ASCII characters are being used in variable names
}
