% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{xpt_check_label_length}
\alias{xpt_check_label_length}
\title{Variable Label Check}
\usage{
xpt_check_label_length(.df)
}
\arguments{
\item{.df}{Data set with labels to check}
}
\description{
Check for variable labels greater than 40 characters in length
}
