% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{xpt_check_var_length}
\alias{xpt_check_var_length}
\title{Variable Length Check}
\usage{
xpt_check_var_length(.data)
}
\arguments{
\item{.data}{Data frame}
}
\value{
Tibble of Variables flagged with lower case
}
\description{
Check to see if variables in data all have length less than or equal to 8
}
