% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xportr.R
\name{xpt_validate}
\alias{xpt_validate}
\title{Validate Dataset Can be Written to xpt}
\usage{
xpt_validate(data)
}
\arguments{
\item{data}{Dataset to be exported as xpt file}
}
\value{
Returns a character vector of failed conditions
}
\description{
Function used to validate dataframes before they are sent to
\code{haven::write_xpt} for writing.
}
