% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_checks.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to a Dataframe}
\usage{
add_labels(.data, ...)
}
\arguments{
\item{.data}{Dataframe with labels to extract}
}
\value{
A vector of variable and labels for a dataframe
}
\description{
Add labels to a Dataframe
}
