% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_spec}
\alias{dataset_spec}
\title{Example Dataset Specification}
\format{
\subsection{\code{dataset_spec}}{

A data frame with 1 row and 9 columns:
\describe{
\item{Dataset}{\if{html}{\out{<chr>}} Dataset}
\item{Description}{\if{html}{\out{<chr>}} Dataset description}
\item{Class}{\if{html}{\out{<chr>}} Dataset class}
\item{Structure}{\if{html}{\out{<lgl>}} Logical, indicating if there's a specific structure}
\item{Purpose}{\if{html}{\out{<chr>}} Purpose of the dataset}
\item{Key, Variables}{\if{html}{\out{<chr>}} Join Key variables in the dataset}
\item{Repeating}{\if{html}{\out{<chr>}} Indicates if the dataset is repeating}
\item{Reference Data}{\if{html}{\out{<lgl>}} Regerence Data}
\item{Comment}{\if{html}{\out{<chr>}} Additional comment}
}
}
}
\usage{
dataset_spec
}
\description{
Example Dataset Specification
}
\keyword{datasets}
