% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\name{minimal_metadata}
\alias{minimal_metadata}
\title{Minimal metadata data frame mock for a ADaM dataset}
\usage{
minimal_metadata(
  dataset = FALSE,
  length = FALSE,
  label = FALSE,
  type = FALSE,
  format = FALSE,
  order = FALSE,
  dataset_name = "adsl",
  var_names = NULL
)
}
\arguments{
\item{dataset}{Flag that indicates that the \code{dataset} column should
be included.}

\item{length}{Flag that indicates that the \code{length} column should
be included.}

\item{label}{Flag that indicates that the \code{label} column should
be included.}

\item{type}{Flag that indicates that the \code{type} column should
be included.}

\item{format}{Flag that indicates that the \code{format} column should
be included.}

\item{order}{Flag that indicates that the \code{order} column should
be included.}

\item{dataset_name}{String with name of domain.}

\item{var_names}{Character vector that defines which variables (rows)
to keep}
}
\value{
A metadata data.frame
}
\description{
Minimal metadata data frame mock for a ADaM dataset
}
\keyword{internal}
