% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\name{minimal_table}
\alias{minimal_table}
\title{Minimal data frame mock of a valid ADaM dataset}
\usage{
minimal_table(n_rows = 3, cols = c("x", "y"))
}
\arguments{
\item{n_rows}{Numeric value that indicates the number of rows of the data
frame}

\item{cols}{Vector of characters that indicates which columns to return.
By default only \code{x} and \code{y} are returned with numeric contents.}
}
\value{
A data.frame mimicking a valid ADaM dataset.
}
\description{
This function is only used in tests.
}
\keyword{internal}
