% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-test.R
\name{multiple_vars_in_spec_helper2}
\alias{multiple_vars_in_spec_helper2}
\title{Test if multiple vars in spec with appropriate}
\usage{
multiple_vars_in_spec_helper2(fun)
}
\description{
Test if multiple vars in spec with appropriate
}
\keyword{internal}
