% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{var_names_log}
\alias{var_names_log}
\title{Utility for Renaming Variables}
\usage{
var_names_log(tidy_names_df, verbose)
}
\arguments{
\item{tidy_names_df}{dataframe}

\item{verbose}{Provides additional messaging for user}
}
\value{
Output to Console
}
\description{
Utility for Renaming Variables
}
