% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{var_ord_msg}
\alias{var_ord_msg}
\title{Utility for Ordering}
\usage{
var_ord_msg(moved_vars, verbose)
}
\arguments{
\item{moved_vars}{Variables moved in the dataset}

\item{verbose}{Provides additional messaging for user}
}
\value{
Output to Console
}
\description{
Utility for Ordering
}
