% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xportr-package.R
\docType{package}
\name{xportr-package}
\alias{xportr}
\alias{xportr-package}
\title{The \code{xportr} package}
\description{
Package Info here
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
