% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xportr-package.R
\docType{package}
\name{xportr-package}
\alias{xportr}
\alias{xportr-package}
\title{The \code{xportr} package}
\description{
Package Info here
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atorus-research/xportr}
  \item Report bugs at \url{https://github.com/atorus-research/xportr/issues}
}

}
\author{
\strong{Maintainer}: Eli Miller \email{Eli.Miller@AtorusResearch.com} (\href{https://orcid.org/0000-0002-2127-9456}{ORCID})

Authors:
\itemize{
  \item Vignesh  Thanikachalam \email{vignesh.x.thanikachalam@gsk.com}
  \item Ben Straub \email{ben.x.straub@gsk.com}
  \item Ross Didenko \email{Ross.Didenko@AtorusResearch.com}
  \item Zelos Zhu \email{Zelos.Zhu@AtorusResearch.com}
}

Other contributors:
\itemize{
  \item Atorus/GSK JPT [copyright holder]
}

}
\keyword{internal}
