% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{xportr_df_label}
\alias{xportr_df_label}
\title{Dataset label}
\usage{
xportr_df_label(.df, datadef, domain = NULL)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}

\item{datadef}{A data frame containing dataset level metadata.}

\item{domain}{A character value to subset the \code{.df}. If \code{NULL}(default), uses
\code{.df} value as a subset condition.}
}
\value{
Data frame with label attributes.
}
\description{
Assigns dataset label from a dataset level metadata to a given data frame.
}
\examples{
adsl <- tibble::tribble(
 ~USUBJID, ~SITEID, ~AGE, ~SEX,
 1001    , 001    , 11  , "M",
 1002    , 002    , 10  , "F",
 1003    , 003    , 14  , "M"
)

datadef <- tibble::tribble(
 ~name , ~label,
 "adsl", "Subject-Level Analysis",
 "adae", "Adverse Events Analysis"
)

adsl <- xportr_df_label(adsl, datadef)
}
\seealso{
\code{\link[=xportr_label]{xportr_label()}}, \code{\link[=xportr_format]{xportr_format()}} and \code{\link[=xportr_length]{xportr_length()}}

Other metadata functions: 
\code{\link{xportr_format}()},
\code{\link{xportr_label}()},
\code{\link{xportr_length}()}
}
\concept{metadata functions}
