% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{xportr_format}
\alias{xportr_format}
\title{SAS format}
\usage{
xportr_format(.df, datadef, domain = NULL)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}

\item{datadef}{A data frame containing variable level metadata.}

\item{domain}{A character value to subset the \code{.df}. If \code{NULL}(default), uses
\code{.df} value as a subset condition.}
}
\value{
Data frame with \code{SASformat} attributes for each variable.
}
\description{
Assigns SAS format from a variable level metadata to a given data frame.
}
\examples{
adsl <- tibble::tribble(
 ~USUBJID, ~BRTHDT,
 1001    , 1,
 1002    , 1,
 1003    , 2
)

datadef <- tibble::tribble(
 ~dataset, ~variable, ~sas_format,
 "adsl"  , "USUBJID", NA,
 "adsl"  , "BRTHDT" , "date9."
 )

adsl <- xportr_format(adsl, datadef)
}
\seealso{
\code{\link[=xportr_label]{xportr_label()}}, \code{\link[=xportr_df_label]{xportr_df_label()}} and \code{\link[=xportr_length]{xportr_length()}}

Other metadata functions: 
\code{\link{xportr_df_label}()},
\code{\link{xportr_label}()},
\code{\link{xportr_length}()}
}
\concept{metadata functions}
