% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{xportr_label}
\alias{xportr_label}
\title{Variable label}
\usage{
xportr_label(.df, datadef, domain = NULL)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}

\item{datadef}{A data frame containing variable level metadata.}

\item{domain}{A character value to subset the \code{.df}. If \code{NULL}(default), uses
\code{.df} value as a subset condition.}
}
\value{
Data frame with label attributes for each variable.
}
\description{
Assigns variable label from a variable level metadata to a given data frame.
}
\examples{
adsl <- tibble::tribble(
 ~USUBJID, ~SITEID, ~AGE, ~SEX,
 1001    , 001    , 11  , "M",
 1002    , 002    , 10  , "F",
 1003    , 003    , 14  , "M"
)

datadef <- tibble::tribble(
 ~dataset, ~variable, ~label,
 "adsl"  , "USUBJID", "Unique Subject Identifier",
 "adsl"  , "SITEID" , "Study Site Identifier",
 "adsl"  , "AGE"    , "Age",
 "adsl"  , "SEX"    , "Sex"
 )

adsl <- xportr_label(adsl, datadef)
}
\seealso{
\code{\link[=xportr_df_label]{xportr_df_label()}}, \code{\link[=xportr_format]{xportr_format()}} and \code{\link[=xportr_length]{xportr_length()}}

Other metadata functions: 
\code{\link{xportr_df_label}()},
\code{\link{xportr_format}()},
\code{\link{xportr_length}()}
}
\concept{metadata functions}
