% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{xportr_length}
\alias{xportr_length}
\title{SAS Length}
\usage{
xportr_length(.df, datadef, domain = NULL)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}

\item{datadef}{A data frame containing variable level metadata.}

\item{domain}{A character value to subset the \code{.df}. If \code{NULL}(default), uses
\code{.df} value as a subset condition.}
}
\value{
Data frame with \code{SASlength} attributes for each variable.
}
\description{
Assigns SAS length from a variable level metadata to a given data frame.
}
\examples{
\dontrun{
adsl <- tibble::tribble(
 ~USUBJID, ~BRTHDT,
 1001    , 1,
 1002    , 1,
 1003    , 2
)

datadef <- tibble::tribble(
 ~dataset, ~variable, ~length,
 "adsl"  , "USUBJID", 10,
 "adsl"  , "BRTHDT" , 8
 )

adsl <- xportr_length(adsl, datadef)
}
}
\seealso{
\code{\link[=xportr_label]{xportr_label()}}, \code{\link[=xportr_df_label]{xportr_df_label()}} and \code{\link[=xportr_format]{xportr_format()}}

Other metadata functions: 
\code{\link{xportr_df_label}()},
\code{\link{xportr_format}()},
\code{\link{xportr_label}()}
}
\concept{metadata functions}
