% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xportr-options.R
\docType{data}
\name{xportr_options}
\alias{xportr_options}
\title{A list with all the supported options of xportr}
\format{
An object of class \code{list} of length 15.
}
\usage{
xportr_options
}
\description{
An internal list with all the supported options of xportr with defaults
}
\keyword{internal}
