% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{xportr_options_list}
\alias{xportr_options_list}
\title{A list with all the supported options of xportr}
\format{
An object of class \code{list} of length 17.
}
\usage{
xportr_options_list
}
\description{
An internal list with all the supported options of xportr with defaults
}
\keyword{internal}
