% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{xportr_order}
\alias{xportr_order}
\title{Order variables of a dataset according to Spec}
\usage{
xportr_order(
  .df,
  metacore = NULL,
  domain = NULL,
  verbose = getOption("xportr.order_verbose", "none")
)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}

\item{metacore}{A data frame containing variable level metadata.}

\item{domain}{A character value to subset the \code{.df}. If \code{NULL}(default), uses
\code{.df} value as a subset condition.}

\item{verbose}{Option for messaging order results}
}
\value{
Dataframe that has been re-ordered according to spec
}
\description{
Order variables of a dataset according to Spec
}
\examples{
adsl <- data.frame(
  BRTHDT = c(1, 1, 2),
  STUDYID = c("mid987650", "mid987650", "mid987650"),
  TRT01A = c("Active", "Active", "Placebo"),
  USUBJID = c(1001, 1002, 1003)
)

metadata <- data.frame(
  dataset = c("adsl", "adsl", "adsl", "adsl"),
  variable = c("STUDYID", "USUBJID", "TRT01A", "BRTHDT"),
  order = 1:4
)

adsl <- xportr_order(adsl, metadata)
}
