% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{xportr_order}
\alias{xportr_order}
\title{Order variables of a dataset according to Spec}
\usage{
xportr_order(
  .df,
  metacore,
  domain = NULL,
  verbose = getOption("xportr.order_verbose", "none")
)
}
\arguments{
\item{.df}{A data frame of CDISC standard.}

\item{metacore}{A data frame containing variable level metadata.}

\item{domain}{A character value to subset the \code{.df}. If \code{NULL}(default), uses
\code{.df} value as a subset condition.}

\item{verbose}{Option for messaging order results}
}
\value{
Dataframe that has been re-ordered according to spec
}
\description{
Order variables of a dataset according to Spec
}
