% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{xportr_write}
\alias{xportr_write}
\title{Write SAS transport file}
\usage{
xportr_write(.df, path, label = NULL)
}
\arguments{
\item{.df}{A data frame to write.}

\item{path}{Path to a file where the data will be written.}

\item{label}{Dataset label. It must be<=40 characters.}
}
\value{
A data frame. \code{write_xport()} returns the input data invisibly.
}
\description{
Writes a local data frame into SAS transport file of version 5. The SAS
transport format is a open format, as is required for submission of the data
to the FDA.
}
\details{
\itemize{
\item Variable and dataset labels are stored in the "label" attribute.
\item SAS length are stored in the "SASlength" attribute.
\item SAS format are stored in the "SASformat" attribute.
\item SAS type are stored in the "SAStype" attribute.
}
}
\examples{
tmp <- tempfile(fileext = ".xpt")
xportr_write(mtcars, tmp, label = "Motor Trend Car Road Tests")
}
