% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{xportr_write}
\alias{xportr_write}
\title{Write xpt v5 transport file}
\usage{
xportr_write(
  .df,
  path,
  metadata = NULL,
  domain = NULL,
  strict_checks = FALSE,
  label = deprecated()
)
}
\arguments{
\item{.df}{A data frame to write.}

\item{path}{Path where transport file will be written. File name sans will be
used as \code{xpt} name.}

\item{metadata}{A data frame containing variable level metadata. See
'Metadata' section for details.}

\item{domain}{Appropriate CDSIC dataset name, e.g. ADAE, DM. Used to subset
the metadata object. If none is passed, then name of the dataset passed as
.df will be used.}

\item{strict_checks}{If TRUE, xpt validation will report errors and not write
out the dataset. If FALSE, xpt validation will report warnings and continue
with writing out the dataset. Defaults to FALSE}

\item{label}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Previously used to to set the Dataset label.
Use the \code{metadata} argument to set the dataset label.}
}
\value{
A data frame. \code{xportr_write()} returns the input data invisibly.
}
\description{
Writes a local data frame into SAS transport file of version 5. The SAS
transport format is an open format, as is required for submission of the data
to the FDA.
}
\details{
\itemize{
\item Variable and dataset labels are stored in the "label" attribute.
\item SAS length are stored in the "SASlength" attribute.
\item SAS format are stored in the "SASformat" attribute.
\item SAS type are stored in the "SAStype" attribute.
}
}
\examples{
adsl <- data.frame(
  SUBL = as.character(123, 456, 789),
  DIFF = c("a", "b", "c"),
  VAL = c("1", "2", "3"),
  PARAM = c("param1", "param2", "param3")
)

var_spec <- data.frame(
  dataset = "adsl",
  label = "Subject-Level Analysis Dataset",
  data_label = "ADSL"
)
xportr_write(adsl,
  path = paste0(tempdir(), "/adsl.xpt"),
  metadata = var_spec,
  strict_checks = FALSE
)

}
